Add export of quiver data to CSV output.
The previous export to CSV for a selected area of water velocity saved
the latitude, longitude, depth and water velocity only. A user requested
that the bearing for the water velocity also be exported. I have now
added that capability. Some implementation details. When a user draws an
area on the map, the water velocity magnitude is displayed as a colour
map on one grid size. If the user selects to display the _bearing_ of
the water velocity (under the Arrows menu), then this bearing (or
quiver) data is displayed on a coarser grid. To implement the new export
capability, I needed both sets of data, the water velocity magnitude and
the water velocity bearing, on the same grid. Therefore, I added the
retaining of the quiver data on the finer grid when the data set is
loaded. For display purposes, the quiver data is still shown on the
_same_ coarser grid (which was probably done originally for aesthetic
reasons because otherwise the image would be too crowded with arrows).
Note the water velocity bearing is actually exported as two components,
water X velocity and water Y velocity, and its export is only triggered
if the user has selected display of that data in the browser.