Reduced the probability of the Failed image being triggered
Made changes so that it takes 2 queries to trigger the failed image to be loaded. This makes it less likely to trigger the race condition that was causing the Faile image (sad computer png) to render over the plot after it loads properly. This is a poor fix as the error is not always triggered when it should be.