Fixes issue #181
The second depth level for the GIOPS Daily dataset is approximately 1.54 m.
In the user interface this value was shown *rounded* to 2 m, while in the CSV
export it was *chopped* to 1 m. This fix applies rounding in both instances.
Furthermore, the units are now also included in the CSV output.