fix for island bug #209
Fix for small island disappearing  bug #209 
This dose not fix the issues for the polar projections yet. 

The code sets the resulation and the size of the islands that are rendered based on the size of the selected area.  
This speeds up the rrender time for large polts form ~50 seconds to ~25 seconds but slows down the small projections from ~5 seconds to ~40 seconds.