Many improvment to area plots
There is now a more complex decision regarding what projection the plot should be. This includes large changes to how plots are created at the poles.
polar projects will now be used less often, only when selected area includ the pole are are very close to it. 
plots that are more than 1/2 of the world will now plot in marcarder projection,
*there are some problems with the plots in the normal projections getting cut off on the edages.