Mostly Working Prototype
The basic functionality and feel of the prototype is there, some bugs need to be fixed, and many additional features need to be implemented before it's ready to deploy.
Additionally, the compare datasets isn't working, so that will be something that requires some thoughts and ideas on the implementation and the ui before being fixed.