Fix Virtual Mooring
The sqlite backend migration changed how datasets are "opened", so most of the plotting classes were broken. The initial merge had fixes for most plot types except for virtual mooring and stick plots.

I've also fixed the timeseries functions in data.py so that we don't construct a range from the timestamp values, and instead have a list of [start, end] from where we find the corresponding indices in the dataset's time dimension.

Given the small amount of test data I have, the results are identical to the same time range on the public site. I'll consider this good to go.