Pare down datasetconfig (#615)
This is in response to a series of errors that were found by clients. Recent code and data changes have introduced instability in the calculation layer. I've fixed and removed several attributes in the config file.

All datasets except GIOPS have been disabled until we can figure out the root cause for the errors.