CLASS4: Replace thredds indexing with FTP indexing (#621)
THREDDS was always slow, but as the number of files increased tomcat began timing out clients trying to generate a pickle. This changes the way CLASS4 files are indexed to use FTP.

@dwayne-hart This introduces a new variable in `oceannavigator.cfg` to be configured.

Pickle files are now generated in <5s, so there is potential for that to happen upon client request, rather than daily.