Add Fvcom to ioc/factory refactoring of open_dataset()
This is a mostly "mechanical" refactoring of Fvcom to correspond to what was
done to the Mercator and Nemo classes. It avoids orphaning the Fvcom class
in this refactoring, but the result has not been subjected to integration
or UI testing because there are presently no FVCOM datasets to work with.