Fix attr access for interp, radius & neighbours
Part of the ioc/factory refactoring in PR#641 moved get_area() from the Data
to the Model class. A NetCDFData instance (which derived from Data) is injected
when instances that derive from Model (Nemo, Mercator, ...) are created. So,
access to the interp, radius & neighbours attrs is now via that injected
instance that is called nc_data.

Also add a unit test for get_data() on the Mercator object.