Clean up attrs access among NetCDFData and Model-derived classes (#682)
* Drop time_variable attr from Model-derived classes

Access it as an attr on NetCDFData.

* Drop timestamps attr from Model-derived classes

Access it as an attr on NetCDFData.

Bonus: This eliminates use of the NetCDFData.meta_only from the NEMO class.

* Drop timestamp_to_time_index() from Model classes

Access it as an attr on NetCDFData.

* Drop _dataset attr from Model-derived classes

Access it as the dataset attr on NetCDFData.

* Drop unused lat/lon attrs from Nemo class

* Drop unused dataset arg in find_nearest_grid_point()

* Fix find_nearest_grid_point() docstring

Drop deleted dataset arg. Missed in 0e54a493f2.