Adds smarter API querying for platforms.
Previously if the user selected tracks inside an area, all points
outside the area were filtered out--this can lead to misleading results
if the platform left the area and later reentered it, it wouldn't show a
discontinunity in the track.
This change makes it so that if a platform passes through the queried
area, the whole track for the time period is returned, even if some
points are outside the area.