Fix shift dimension selection in unstaggered_speed()
Calculated variable equation parser won't let us passing dimensions or their
names as strings. So, we have to assume that the dimension order of the
velocity component arrays is (t, depth, y, x) or (t, y, x). That means that
we can pick out the dimensions that we need to shift along from the dims attr
using -2 and -1 indices.