Add SalishSeaCast currents dataset config, colour maps & dataset handling (#709)
* Fix typo in unit test

This doesn't make any difference to the test outcome, but the dataset key
and the netCDF file that the test operates on should be consistent so that
we don't confuse our future selves.

* Enabled skipped test_enter_grid_angle_file()

This commit adds a small-ish test data grid angle file that is consistent
with the exiting testdata/salishseacast_ssh_test.nc so that we can enable
testing of the grid angle file branch of the context manager entry.

* Use xarray.open_dataset() not open_mfdataset()

Grid angle files are, by specification, single files, not collections of files,
so there is no need to use open_mfdataset(). Doing so is perhaps slower.

* Simplify merging of geo-refs and fields datasets

* Fix typo in SalishSeaCast bio dataset help blurb

* Fix typos in datasetconfig.json test fixture

re: lat_var_key & lon_var_key values for SalishSeaCast dataset fixture
This is for correctness; no impact on pass/fail of any existing tests.

* Don't lowercase dataset variable config key

Doing so precludes the use of data sources that use mixed case variable names.
With the introduction of ingesting datasets from ERDDAP/OPeNDAP data sources
where we can't control the variables names, we can't support restructing them
to lowercase only.

* Handle list of dataset URLs

This facilitates merging variables from multiple OPeNDAP/ERDDAP
data sources into a single, lazy-loaded NetCDFData instance.

* Add staggered grid current calculation function

Calculate the speed of seawater current from u and v velocity component
array that are on the u and v points of an Akawara-C staggered grid;
see https://en.wikipedia.org/wiki/Arakawa_grids

* Add SalishSeaCast currents dataset config & colormaps

* Fix shift dimension selection in unstaggered_speed()

Calculated variable equation parser won't let us passing dimensions or their
names as strings. So, we have to assume that the dimension order of the
velocity component arrays is (t, depth, y, x) or (t, y, x). That means that
we can pick out the dimensions that we need to shift along from the dims attr
using -2 and -1 indices.