Updated the [...]/Ocean-Data-Map-Project/oceannavigator/configs (#758)
hash so that when people go to install or update their instance
of the Ocean Navigator. They will receive the appropriate
configuration files.

Also, I updated the indexing-tool's conda specification file as
the version of openssl it was refering to was no longer
available from either conda-forge or alternate channel we
currently use.