Change Show Location checkbox default to unchecked
This speeds up rendering of the Points and Class4 UI overlay contents by 2-3
seconds at the expense of not defaulting to show a fairly uninformative map with
a point on it. The user is free to check Show Location if they want to see the
map.

The intent of this change is to provide a temporary UX speed improvement until
frontend work is undertaken to update the Points and Class4 UI.