Update runserver.sh
Statically set the value to the WORKER_THREADS variable to 5. Otherwise if we allow the system to use the rule of thumb (2*NUMBER_WORKERS) to set the number of WORKER_THREADS we will run out of memory resources within the LXD/LXC containers.

Also, to cut down on the virutal RAM usage the "--worker-tmp-dir /dev/shm" option for gunicorn was removed. The default behavior is for gunicorn to use the /tmp directory for writing temporary files.