Add CHS baselayer (#804)
* Add CHS baselayer

* Fix dimensions do not exist case in issue #769 (#780)

Implement solution by @nsoontie that adds properties to NetCDFData class to
facilitate finding y and x dimension DataArrays in Datasets with a variety of
dimensions via set intersections.

* Add Sentry integration for backend via Flask app (#776)

* Add Sentry integration for backend via Flask pp

* Add API endpoint to test Sentry on the backend

Hitting /api/test-sentry results in a transaction event being captured by
Sentry.

* Changing code owner from Nabil to myself. (#782)

* Changing code owner from Nabil to myself.

* Adding Vanessa into the CODEOWNERS list.

Co-authored-by: Dwayne Hart <dwayne.hart@dfo-mpo.gc.ca>

* Add possible Sentry environment name from envvar (#788)

* Drop unused `pint.UnitRegistry` in `Mercator.depths` & `Nemo.depths` properties (#791)

* Drop use of pint.UnitRegistry in depths property

UnitRegistry takes ~200ms per API call to instantiate but its features are
never used for the Nemo or Mercator depths.

* Make depths property immutability setting consistent

All other places in the code base where the writeable flag of a Numpy array is
set use setflags(write=...), so do the same in Mercator.depths.

* Delete trailing whitespace

* Only use open_mfdataset() when there is >1 dataset (#789)

There is a measurable performance cost to using xarray.open_mfdataset() with
chunking instead of xarray.open_dataset() without chunking when we are only
opening a single dataset.

* Get Sentry traces_sample_rate value from envvar (#792)

* Delete unused UnitRegistry imports (#793)

Codebeat is being a nussence.

* Created a new conda environment called cdo-tools which contains the (#783)

Unidata cdo aplications.

Installed lftp in the devops conda environment and created an
updated spec file.

Updated the navigator spec file due in part to the following changes;
 * Installed a newer version of gunicorn (20.0.4)
 * Installed sentry-sdk and blinker for the API debugging utility
 * Installed defopt which is needed for the Observations project
 * Installed geojson as this is needed by Nabil Miri for development
   work he is undertaking.

Updated the hash signature of the configs submodule.

* Fix for issue #796 re: class4/models/{class4_id} API endpoint (#797)

* Add unit test that demonstrates issue #796

* Add year to class4 file path to search for profiles

This reflects the /data/class4/yyyy/yyyymmdd/ file storage scheme that is being
used.

Also parse the yyyymmdd section out of the class4 id into variables to improve
code readability.

* Refactor list_class4_models() to use pathlib

* Create codeql-analysis.yml (#794)

Automated code vulnerability analysis for JS and Python.

Co-authored-by: Dwayne Hart <44408505+dwayne-hart@users.noreply.github.com>

* Zarr data access enabled (#779)

* Zarr feature source code update

* source code updated & new test cases added

* Code cleanup and unittest code modification

* updating with latest changes in master

* Update codeql-analysis.yml

Corrected line 16.

* Update the hash for the configs submodule. (#800)

Co-authored-by: Dwayne Hart <dwayne@gmail.com>

* Navigator specfile 20201210 (#802)

* Creating a new spec file for the navigator toolchain as zarr is
a required component for Zarr data access enabled #779.

* Ensure that the hash for oceannavigator/configs refers to the off-site
branch.

Co-authored-by: Dwayne Hart <dwayne.hart@dfo-mpo.gc.ca>

* zarr unittest issue fix (#803)

* Implement Depth Excess acoustic variable (#753)

* Add bathymetry_file_url to datasetconfig

* Implement Depth Excess acoustic variable

Co-authored-by: Dwayne Hart <44408505+dwayne-hart@users.noreply.github.com>

* Change Class 4 profiles legend to show GIOPS model label from the outset (#806)

* Chg class4 initial profiles model legend to GIOPS

* Update string formatting to f-strings

* Updating the configs submodule hash so that when a person runs the (#807)

git submodule update --init --recursive command, The configs
directory will have the a link to

    commit 5c7e47638996738713fea84394f969156f747301

* Fix for issue #798 re: FileNotFoundError in class4 forecasts API request (#799)

* Add unit test that demonstrates issue #798

Also fix typos in TestListClass4Models that were missed in PR#797.

* Fix parsing of class4 id to dataset URL

Strip class4 index number form end of class4 id.

Also:
* Add type annotations to function signature
* Add docstring
* Change returned variable name to result for consistency with
  list_class4_models()

 re: issue #798

* Fix AttributeError re: class4 modeljuld units

Opening the dataset with decode_times=False is key because we do the date/time
decoding explicitly.

* Update README.md

Modified the Local Installation instructions.

* Fix UnboundLocalError netCDFData.dimensions (#811)

* Increase test coverage of NetCDFData.dimensions

Note that these tests use NetCDFData() constructor rather than
NetCDFData.__enter__() context manager because of the use case in
data.open_dataset() where the dataset dimensions are accessed in order to
decide what flavour of model dataset object to create.

* Add unit test & type hint for list of dataset URLs

* Fix sqlite3.OperationalError passing silently

Add empty dimension list to prevent UnboundLocalError that hides the
OperationalError. The latter is likely indicative of a deeper infrastructure
problem.

re: https://sentry.io/organizations/dfo-ocean-navigator/issues/2134767923

* Clean up trailing whitespace

* Update class4 index pickle generator script (#809)

* Update generate_class4_list script

* Use pathlib.Path because Python>=3.6
* Use set and list comprehensions because modern Python
* Sort the class4 files dict in reverse date order because MapToolbar.jsx uses
the first element to set the month displayed and we want that to be the most
recently indexed class4 files.
* Rename data.class4.__list_class4_files_slowly() to
  data.class4._list_class4_files_slowly() because leading-single-underscore
  "internal use" convention is sufficient; leading-double-underscore name
  mangling just makes life difficult for IDE users
* Use generate_class4_list.list_class4_files() in
  data.class4._list_class4_files_slowly() to eliminate duplicated code

* Drop unnecessary `#!/usr/bin/env python`

* Ensure init of class4 pickle lock mgmt vars

* Drop bare Exception re: pickle.dump()

It's unclear what error condition this is supposed to catch and bare Exceptions
are bad form. We can add specific exception handling back if/when we figure out
what this try/except was protecting against.

* Change to use f-strings

* Update module docstring

* Update runserver.sh

Statically set the value to the WORKER_THREADS variable to 5. Otherwise if we allow the system to use the rule of thumb (2*NUMBER_WORKERS) to set the number of WORKER_THREADS we will run out of memory resources within the LXD/LXC containers.

Also, to cut down on the virutal RAM usage the "--worker-tmp-dir /dev/shm" option for gunicorn was removed. The default behavior is for gunicorn to use the /tmp directory for writing temporary files.

* Nuke stats tab (#812)

* Bumping the configs has tag to 42c8b17 for the off-site branch. (#813)

This allows consumers of the ONAV project a way to quickly
deply their own instance.

Co-authored-by: Dwayne Hart <dwayne.hart@dfo-mpo.gc.ca>

* s

Co-authored-by: Doug Latornell <doug.latornell@43ravens.ca>
Co-authored-by: Dwayne Hart <44408505+dwayne-hart@users.noreply.github.com>
Co-authored-by: Dwayne Hart <dwayne.hart@dfo-mpo.gc.ca>
Co-authored-by: Lavan Thaya <46497911+CubeCode93@users.noreply.github.com>
Co-authored-by: Dwayne Hart <dwayne@gmail.com>